fix(mla-tq): cast rotate_query output back to bf16 in TurboQuantMLABackend
Per code review round 2 of the kernel path: tq_config.rotate_query
returns fp32 regardless of input dtype (CUDA helper convention).
Passing fp32 Q_nope into the Triton kernel doubled register pressure
versus bf16 and introduced a needless precision promotion.

Casting back to bf16 keeps the kernel's expected input dtype and
matches the existing MHA TurboQuant backend's pattern (which also
keeps Q in the KV's dtype after rotation).

Note: this does NOT fix the remaining kernel correctness bug — v1
still outputs ~10x the expected magnitude with cos ≈ 0.14 on the
synthetic correctness test. Debugging continues in a subsequent
session. See Design - Stage C Phase 1 findings.md "Phase 2 — v1
kernel implementation + correctness test" for the repro details.

+++ python/sglang/srt/layers/attention/flashmla_backend.py
@@ -753,7 +753,10 @@ class TurboQuantMLABackend(FlashMLABackend):
         # This is the move that activates the orthogonality trick, letting the
         # kernel dot rotated Q against rotated K without needing per-row
         # inverse Hadamard.
-        q_nope_rot = self._tq_config.rotate_query(q_nope)
+        # rotate_query returns fp32 regardless of input dtype (CUDA kernel
+        # convention). Cast back to bf16 to match kernel's expected dtype
+        # and keep register pressure bounded.
+        q_nope_rot = self._tq_config.rotate_query(q_nope).to(q_nope.dtype)
 
         # Pool references (current layer). Kernel reads these directly.
         layer_id_rel = layer.layer_id - self._tq_pool.start_layer
